Cover the card binding, share the drawer-content harness, make the field required
